+++ rotkehlchen/chain/optimism/modules/velodrome/balances.py
@@ -3,6 +3,8 @@ from typing import TYPE_CHECKING
 from rotkehlchen.chain.evm.decoding.velodrome.balances import VelodromeLikeBalances
 from rotkehlchen.chain.evm.decoding.velodrome.constants import CPT_VELODROME
 
+from .constants import A_VELO, VOTING_ESCROW_CONTRACT_ADDRESS
+
 if TYPE_CHECKING:
     from rotkehlchen.chain.optimism.decoding.decoder import OptimismTransactionDecoder
     from rotkehlchen.chain.optimism.node_inquirer import OptimismInquirer
@@ -18,4 +20,6 @@ class VelodromeBalances(VelodromeLikeBalances):
             evm_inquirer=evm_inquirer,
             tx_decoder=tx_decoder,
             counterparty=CPT_VELODROME,
+            voting_escrow_address=VOTING_ESCROW_CONTRACT_ADDRESS,
+            protocol_token=A_VELO.resolve_to_evm_token(),
         )